fixes are complete and ready to be tested live

+++ backend/db/migrations/20231001165708-create-event.js
@@ -49,7 +49,7 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL(6,2)
       },
       startDate: {
         type: Sequelize.DATE,
